+++ app/src/main/java/com/example/androidsample/MainActivity.kt
@@ -1,0 +1,169 @@
+/*
+ * Copyright 2021 HyperDevs
+ *
+ * Copyright 2020 BQ
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.androidsample
+
+import android.os.Bundle
+import android.view.View
+import android.widget.ProgressBar
+import android.widget.TextView
+import android.widget.Toast
+import androidx.activity.viewModels
+import androidx.lifecycle.SavedStateHandle
+import kotlinx.coroutines.*
+import mini.*
+import mini.android.FluxActivity
+import mini.android.FluxStoreViewModel
+import kotlinx.coroutines.flow.onEach
+import java.io.Serializable
+
+val dispatcher = Dispatcher()
+
+data class State(
+        val text: String = "0",
+        val loading: Boolean = false
+) : Serializable
+
+@Action
+data class SetLoadingAction(val loading: Boolean)
+
+@Action
+data class SetTextAction(val text: String)
+
+@Action
+interface AnalyticsAction
+
+@Action
+class LongUseCaseAction(val userName: String) : AnalyticsAction, SuspendingAction
+
+
+/**
+ * Use any name you like for suspending actions, or use reducer
+ */
+typealias UseCase = Reducer
+
+class MainStore : Store<State>() {
+
+    init {
+        Mini.link(dispatcher, this).track()
+    }
+
+    @Reducer
+    fun handleLoading(state: State, action: SetLoadingAction): State {
+        return state.copy(loading = action.loading)
+    }
+
+
+    @Reducer
+    fun handleSetTextAction(state: State, action: SetTextAction): State {
+        return state.copy(text = action.text)
+    }
+
+    @Reducer
+    fun handleAnalyticsAction(action: AnalyticsAction) {
+        //Log to analytics
+    }
+
+    @Reducer
+    fun handleAnyAction(action: Any) {
+        //Log to analytics
+    }
+
+    @UseCase
+    suspend fun useCase(s: LongUseCaseAction) {
+        if (state.loading) return
+        dispatcher.dispatch(SetLoadingAction(true))
+        dispatcher.dispatch(SetTextAction("Loading from network..."))
+        delay(5000)
+        dispatcher.dispatch(SetTextAction("Hello From UseCase"))
+        dispatcher.dispatch(SetLoadingAction(false))
+    }
+}
+
+class MainViewModelReducer : NestedStateContainer<State>() {
+
+    @Reducer
+    fun handleLoading(state: State, action: SetLoadingAction): State {
+        return state.copy(loading = action.loading)
+    }
+
+    @Reducer
+    fun handleSetTextAction(state: State, action: SetTextAction): State {
+        return state.copy(text = action.text)
+    }
+}
+
+class MainStoreViewModel(savedStateHandle: SavedStateHandle) : FluxStoreViewModel<State>(savedStateHandle) {
+    private val reducerSlice = MainViewModelReducer().apply { parent = this }
+
+    init {
+        Mini.link(dispatcher, listOf(this, reducerSlice)).track()
+    }
+
+    override fun saveState(state: State, handle: SavedStateHandle) {
+        println("State saved")
+        handle.set("state", state)
+    }
+
+    override fun restoreState(handle: SavedStateHandle): State? {
+        val restored = handle.get<State>("state")
+        println("State restored $restored")
+        return restored
+    }
+
+    @UseCase
+    suspend fun useCase(action: LongUseCaseAction) {
+        if (state.loading) return
+        dispatcher.dispatch(SetLoadingAction(true))
+        delay(2000)
+        dispatcher.dispatch(SetTextAction("${state.text.toInt() + 1}"))
+        dispatcher.dispatch(SetLoadingAction(false))
+    }
+}
+
+class MainActivity : FluxActivity() {
+
+    lateinit var textView: TextView
+    lateinit var progressBar: ProgressBar
+    private val vm: MainStoreViewModel by viewModels()
+
+    override suspend fun whenCreated(savedInstanceState: Bundle?) {
+        setContentView(R.layout.main_activity)
+        textView = findViewById(R.id.textView)
+        progressBar = findViewById(R.id.progressBar)
+
+        textView.setOnClickListener {
+            launch {
+                dispatcher.dispatch(LongUseCaseAction("Pablo"))
+                // Decide on the state after usecase is done
+                // I won't run until use case is done
+            }
+        }
+
+        vm.flow().onEach {
+            textView.text = it.toString()
+            progressBar.visibility = if (it.loading) View.VISIBLE else View.INVISIBLE
+        }.launchInLifecycleScope()
+
+        vm.flow()
+                .select { it.loading }
+                .onEachDisable {
+                    Toast.makeText(this@MainActivity, "Finished loading", Toast.LENGTH_SHORT).show()
+                }.launchInLifecycleScope()
+    }
+}