+++ mini-android/src/main/java/mini/android/FluxViewModel.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2021 HyperDevs
+ *
+ * Copyright 2020 BQ
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package mini.android
+
+import androidx.annotation.CallSuper
+import androidx.lifecycle.ViewModel
+import mini.CloseableTracker
+import mini.DefaultCloseableTracker
+
+abstract class FluxViewModel : ViewModel(),
+    CloseableTracker by DefaultCloseableTracker() {
+
+    @CallSuper
+    override fun onCleared() {
+        super.onCleared()
+        close()
+    }
+}