+++ mini-common/src/main/java/mini/Resource.kt
@@ -31,10 +31,10 @@ open class Resource<out T> @PublishedApi internal constructor(val value: Any?) {
     val isEmpty: Boolean get() = value is Empty
     val isFailure: Boolean get() = value is Failure
     val isLoading: Boolean get() = value is Loading<*>
-
     val isTerminal: Boolean get() = isSuccess || isFailure
+    val isIdle: Boolean get() = isEmpty
 
-    internal object Empty {
+    internal class Empty {
         override fun toString(): String = "Empty()"
     }
 
@@ -64,17 +64,21 @@ open class Resource<out T> @PublishedApi internal constructor(val value: Any?) {
             else -> null
         }
 
+    @Throws(NullPointerException::class)
+    fun get() : T = getOrNull()!!
+
     companion object {
-        fun <T> success(value: T): Resource<T> = Resource(value)
+        fun <T> success(value: T? = null): Resource<T> = Resource(value)
         fun <T> failure(exception: Throwable? = null): Resource<T> = Resource(Failure(exception))
         fun <T> loading(value: T? = null): Resource<T> = Resource(Loading(value))
-        fun <T> empty(): Resource<T> = Resource(Empty)
+        fun <T> empty(): Resource<T> = Resource(Empty())
+        fun <T> idle(): Resource<T> = empty()
     }
 
     override fun toString(): String {
         return if (isSuccess) "Success($value)" else value.toString()
     }
-
+    
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
@@ -90,42 +94,9 @@ open class Resource<out T> @PublishedApi internal constructor(val value: Any?) {
 }
 
 /**
- * A resource that abstracts asynchronous operation but with idle
- * state instead of empty.
- *
- * It accepts temporary metadata objects as value.
+ * An alias for a empty resource.
  */
-open class TypedTask<T> @PublishedApi internal constructor(value: Any?) : Resource<T>(value) {
-    val isIdle: Boolean get() = isEmpty
-
-    companion object {
-        fun <T> success(value: T): TypedTask<T> = TypedTask(value)
-        fun <T> idle(): TypedTask<T> = TypedTask(Empty)
-        fun <T> loading(value: T? = null): TypedTask<T> = TypedTask(Loading(value))
-        fun <T> failure(exception: Throwable? = null): TypedTask<T> = TypedTask(Failure(exception))
-    }
-
-    override fun toString(): String {
-        return when {
-            isSuccess -> "Success"
-            isIdle -> "Idle"
-            else -> value.toString()
-        }
-    }
-}
-
-/**
- * An empty resource that just abstracts asynchronous operation but with idle
- * state instead of empty.
- */
-class Task(value: Any?) : TypedTask<Unit>(value) {
-    companion object {
-        fun success(): Task = Task(Unit)
-        fun idle(): Task = Task(Empty)
-        fun loading(): Task = Task(Loading<Unit>())
-        fun failure(exception: Throwable? = null): Task = Task(Failure(exception))
-    }
-}
+typealias Task = Resource<Nothing?>
 
 inline fun <T> Resource<T>.onSuccess(crossinline action: (data: T) -> Unit): Resource<T> {
     if (isSuccess) action(value as T)
@@ -142,39 +113,37 @@ inline fun <T> Resource<T>.onLoading(crossinline action: (data: T?) -> Unit): Re
     return this
 }
 
-inline fun <T> TypedTask<T>.onIdle(crossinline action: () -> Unit): TypedTask<T> {
-    if (isEmpty) action()
-    return this
-}
-
 inline fun <T> Resource<T>.onEmpty(crossinline action: () -> Unit): Resource<T> {
     if (isEmpty) action()
     return this
 }
+
+/** Alias of [onEmpty] for Task */
+inline fun Task.onIdle(crossinline action: () -> Unit) = onEmpty(action)
 
 inline fun <T, R> Resource<T>.map(crossinline transform: (data: T) -> R): Resource<R> {
     if (isSuccess) return Resource.success(transform(value as T))
     return Resource(value)
 }
 
-/** All resources completed, whether they're in success or failure state. */
-fun <T> Iterable<Resource<T>>.allTerminal(): Boolean {
-    return this.all { it.isTerminal }
-}
-
-/** All resources succeeded. */
-fun <T> Iterable<Resource<T>>.allSuccesful(): Boolean {
+/** All tasks succeeded. */
+fun <T> Iterable<Resource<T>>.allSuccessful(): Boolean {
     return this.all { it.isSuccess }
 }
 
-/** Any resources failed. */
+/** Any tasks failed. */
 fun <T> Iterable<Resource<T>>.anyFailure(): Boolean {
     return this.any { it.isFailure }
 }
 
-/** Any resource is loading. */
+/** Any task is running. */
 fun <T> Iterable<Resource<T>>.anyLoading(): Boolean {
     return this.any { it.isLoading }
+}
+
+/** All resources completed, whether they're in success or failure state. */
+fun <T> Iterable<Resource<T>>.allTerminal(): Boolean {
+    return this.all { it.isTerminal }
 }
 
 /** Any resource empty */
@@ -186,7 +155,7 @@ fun <T> Iterable<Resource<T>>.onAllTerminal(fn: () -> Unit): Iterable<Resource<T
 }
 
 fun <T> Iterable<Resource<T>>.onAllSuccessful(fn: () -> Unit): Iterable<Resource<T>> {
-    if (this.allSuccesful()) fn()
+    if (this.allSuccessful()) fn()
     return this
 }
 
@@ -210,3 +179,4 @@ fun Iterable<Task>.onAnyIdle(fn: () -> Unit): Iterable<Task> = onAnyEmpty(fn).ma
 /** Returns the first exception that can be found in a list of resources, null if it can't find any */
 fun <T> Iterable<Resource<T>>.firstExceptionOrNull() : Throwable? =
     this.firstOrNull { it.isFailure && it.exceptionOrNull() != null }?.exceptionOrNull()
+