+++ mini-common/src/test/kotlin/mini/TestDispatcher.kt
@@ -21,8 +21,10 @@ package mini
 import kotlin.reflect.KClass
 import kotlin.reflect.jvm.jvmErasure
 
-fun TestDispatcher(): Dispatcher {
-    return Dispatcher(newReflectiveMap())
+fun newTestDispatcher(): Dispatcher {
+    return Dispatcher().apply {
+        actionTypeMap = newReflectiveMap()
+    }
 }
 
 private fun reflectActionTypes(type: KClass<*>, depth: Int = 0): List<ReflectedType> {
@@ -48,7 +50,7 @@ private fun newReflectiveMap(): Map<KClass<*>, List<KClass<*>>> {
         override fun containsKey(key: KClass<*>): Boolean = map.containsKey(key)
         override fun containsValue(value: List<KClass<*>>): Boolean = map.containsValue(value)
         override fun isEmpty(): Boolean = map.isEmpty()
-        override fun get(key: KClass<*>): List<KClass<*>>? {
+        override fun get(key: KClass<*>): List<KClass<*>> {
             return map.getOrPut(key) {
                 reflectActionTypes(key)
                     .asSequence()